:art: use annotations on consolidator module instead of expliciit declaration in xml
