+++ engine/asqatasun-engine/src/main/java/org/asqatasun/service/command/factory/AuditCommandFactoryImpl.java
@@ -27,6 +27,7 @@ import java.util.Set;
 
 import org.apache.log4j.Logger;
 import org.asqatasun.analyser.AnalyserService;
+import org.asqatasun.consolidator.ConsolidatorService;
 import org.asqatasun.contentadapter.AdaptationListener;
 import org.asqatasun.entity.parameterization.Parameter;
 import org.asqatasun.entity.service.audit.AuditDataService;
