+++ engine/asqatasun-engine/src/test/java/org/asqatasun/service/command/AuditCommandTestCase.java
@@ -39,7 +39,7 @@ import org.asqatasun.entity.service.parameterization.ParameterDataService;
 import org.asqatasun.entity.service.reference.TestDataService;
 import org.asqatasun.entity.service.subject.WebResourceDataService;
 import org.asqatasun.analyser.AnalyserService;
-import org.asqatasun.service.ConsolidatorService;
+import org.asqatasun.consolidator.ConsolidatorService;
 import org.asqatasun.service.ContentAdapterService;
 import org.asqatasun.service.ProcessorService;
 import org.asqatasun.util.http.HttpRequestHandler;
