--- engine/asqatasun-resources/src/main/resources/conf/context/beans-analyser.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<beans xmlns="http://www.springframework.org/schema/beans"
-       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-       xsi:schemaLocation="
-       http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-3.0.xsd">
-
-    <bean id="analyserFactory" class="org.asqatasun.analyser.AnalyserFactoryImpl"/>
-
-    <bean id="analyserService" class="org.asqatasun.service.AnalyserServiceImpl">
-        <property name="analyserFactory" ref="analyserFactory"/>
-    </bean>
-
-</beans>
