+++ testing-tools/asqatasun-rules-testing-tools/src/main/resources/context/application-context.xml
@@ -1,6 +1,10 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <beans xmlns="http://www.springframework.org/schema/beans" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-       xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-3.0.xsd">
+       xmlns:context="http://www.springframework.org/schema/context"
+       xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-3.0.xsd http://www.springframework.org/schema/context https://www.springframework.org/schema/context/spring-context.xsd">
+
+
+    <context:component-scan base-package="org.asqatasun.consolidator"/>
 
     <import resource="classpath:context/beans-factories.xml"/>
     <import resource="classpath:context/beans-persistence.xml"/>
@@ -8,7 +12,6 @@
     <import resource="classpath:context/beans-contentloader.xml"/>
     <import resource="classpath:context/beans-contentadapter.xml"/>
     <import resource="classpath:context/beans-processor.xml"/>
-    <import resource="classpath:context/beans-consolidator.xml"/>
 
     <import resource="classpath:context/beans-ruleimplementationloader.xml"/>
     <import resource="classpath:context/beans-nomenclatureloader.xml"/>
