--- testing-tools/asqatasun-rules-testing-tools/src/main/resources/context/beans-consolidator.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<beans xmlns="http://www.springframework.org/schema/beans"
-       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-       xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-3.0.xsd">
-
-    <bean id="consolidatorFactory" class="org.asqatasun.consolidator.ConsolidatorFactoryImpl"/>
-    
-    <bean id="consolidatorService" class="org.asqatasun.service.ConsolidatorServiceImpl"/>
-
-</beans>
